disallow __proto__ and constructor in object fields
__proto__ and constructor collide with the js prototypes. If we allow setting
those from user data we allow somebody to set the prototype of the constructed
object which might lead to really hard to figure out errors.

This becomes an issue when dealing with value classes where we actually need to
have the prototypes pointing to the value class instance and thus cannot just
do prototypeless objects with Object.create(null)

this is a bit of hack as there might well be valid uses of at least
'constructor' in some schema.  Better way would be to stop using classes and
change the isN checks to use some other way to distinguish objects like known
symbols or weakmaps. However as we rely in all places on instanceOf this is not
particularly easy. Or we could override instanceOf symbol to do this. All of
these would require more coding though so not sure if that is now warranted.