fix: optional query params failing when no query params are provided
In openapi you can declare a query param that is optional like so:
```
- name: foo
  in: query
  required: false
  schema:
    type: boolean
```

However, now if you try to call the endpoint _without_ providing the
query parameter you will get an error:
"invalid request query : expected an object, but got \"null\" instead."

This happens because we are tossing the empty query params object away
in this `voidify` function. This commit removes the voidification of
query parameters, which fixes the issue.